+++ CHANGES.rst
@@ -3,7 +3,9 @@ Changelog
 
 2.0.0 (unreleased)
 ------------------
+
 - #1733 Allow results interpretation in sample received state
+- #1732 Readonly Transactions
 - #1731 Remove `notifyModified` method from analyses